requirements.txt
flask
pillow
